requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==23.0.0

jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1

langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchain-pinecone==0.1.3

google-generativeai==0.7.2

pinecone==5.3.1

sentence-transformers==3.0.1
huggingface-hub==0.24.6

numpy==1.26.4
pypdf==4.3.1
requests==2.32.3
PyYAML==6.0.3